+++ contracts/rewards/Vesting02.sol
@@ -24,7 +24,7 @@ contract Vesting02 is ERC721URIStorageUpgradeable, OwnableUpgradeable {
 
     struct Vest {
         address pool;
-        uint256 depositID;
+        uint64 depositID;
         uint256 lastUpdateTimestamp;
         uint256 accumulatedAmount;
         uint256 withdrawnAmount;
@@ -41,7 +41,7 @@ contract Vesting02 is ERC721URIStorageUpgradeable, OwnableUpgradeable {
     event ECreateVest(
         address indexed to,
         address indexed pool,
-        uint256 depositID,
+        uint64 depositID,
         uint256 vestAmountPerStablecoinPerSecond
     );
     event EUpdateVest(uint256 indexed vestID);
@@ -76,7 +76,7 @@ contract Vesting02 is ERC721URIStorageUpgradeable, OwnableUpgradeable {
     function createVestForDeposit(
         address to,
         address pool,
-        uint256 depositID,
+        uint64 depositID,
         uint256 vestAmountPerStablecoinPerSecond
     ) external returns (uint256 vestID) {
         require(
@@ -105,7 +105,7 @@ contract Vesting02 is ERC721URIStorageUpgradeable, OwnableUpgradeable {
     }
 
     function updateVestForDeposit(
-        uint256 depositID,
+        uint64 depositID,
         uint256 currentDepositAmount,
         uint256 depositAmount,
         uint256 vestAmountPerStablecoinPerSecond
